Moved Services, Dto and Exceptions from bundle to application
