+++ src/OpenConext/EngineBlock/Exception/RuntimeException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace OpenConext\EngineBlockBundle\Exception;
+namespace OpenConext\EngineBlock\Exception;
 
 use RuntimeException as CoreRuntimeException;
