+++ src/OpenConext/EngineBlockBundle/Controller/Api/MetadataController.php
@@ -5,8 +5,8 @@ namespace OpenConext\EngineBlockBundle\Controller\Api;
 use OpenConext\EngineBlockBundle\Http\Exception\ApiAccessDeniedHttpException;
 use OpenConext\EngineBlockBundle\Http\Exception\BadApiRequestHttpException;
 use OpenConext\EngineBlockBundle\Http\Response\JsonResponse;
-use OpenConext\EngineBlockBundle\Service\FeaturesService;
-use OpenConext\EngineBlockBundle\Service\MetadataService;
+use OpenConext\EngineBlock\Service\FeaturesService;
+use OpenConext\EngineBlock\Service\MetadataService;
 use OpenConext\EngineBlockBundle\Http\Response\JsonHelper;
 use OpenConext\Value\Exception\InvalidArgumentException;
 use OpenConext\Value\Saml\EntityId;
